Conductor, fish and indicator classes
Conductor: calculates necesary values to sync our notes (fish) with the song, spawn the notes.

Fish: it has a coroutine that moves the fish depending on the time in beats calculated by de Conductor class

Indicator: Just print int console the Conductor's time in beats when the fish reach the indicator.