Pause Menu Somewhat Functional and Broken
Here are the things I changed.

For Fish.cs:
- In the move method, I added in a while loop to that will loop forever when the game is paused. This is to stop all movement of the note.

For Conductor.cs
- I moved the code in the Update() method to the FixedUpdate() method. This is because when Time.timescale(0), the FixedUpdate() method will not get called unlike the Update() method. Also, it seems to be better to have all physics in the FixedUpdate() method anyways.
- Added in a GetMusicSource Method.
- Added other methods/ variables that will need to be changed.

Made changes to GameUIController and added in a test class for Conductor class. Currently I am not testing any methods in the Conductor class since I was having trouble with assemblies breaking Imports for the Input System package. I will have to look into that at some point.