Fixed bug where player can get points when paused
The EventListener in the Indicator.cs class continued to work even when the game is paused. I added in a if check that will return when the game is paused.

For the Fish.cs class, I added in a isPaused check in the Escape() method. Something I noticed is since the Move and escape coroutines track where they are based on the song/beat time, the isPaused checks are not needed to stop the fish's/note's movement. I am not sure if we gain any speed from having the coroutines wait until the game is not paused or not. I left it in for now and will get a second opinion in code review.